some error fixed

+++ Controllers/Admin.js
@@ -2,14 +2,20 @@ import Agent from "../Models/Agent.js";
 import Property from "../Models/Property.js";
 import Owner from "../Models/Owner.js";
 import User from "../Models/User.js";
+import AdBanner from "../Models/AdBanner.js";
+import PropertyAd from "../Models/PropertyAd.js";
+import SoldProperties from "../Models/SoldProperties.js";
 //admin dashboard get total for all
 import AgentCompany from "../Models/AgentCompany.js";
 export const adminDashboard = async (req, res) => {
   try {
     const properties = await Property.find().count();
-    const soldProperties = await Property.find({isSoldOut:true}).count();
-    const rentedProperties = await Property.find({isRented:true}).count();
-    const featuredProperties = await Property.find({isFeatured:true}).count();
+    const soldProperties = await Property.find({ isSoldOut: true }).count();
+    const rentedProperties = await Property.find({ isRented: true }).count();
+    const featuredProperties = await Property.find({
+      isFeatured: true,
+    }).count();
+    const banners = await AdBanner.find({ isAccepted: true }).count();
     const agentCompanies = await AgentCompany.find().count();
     const agents = await Agent.find().count();
     const users = await User.find().count();
@@ -20,9 +26,10 @@ export const adminDashboard = async (req, res) => {
         totalCompanies: agentCompanies,
         totalAgent: agents,
         totalUsers: users,
-        totalRentedProperties:rentedProperties,
-        totalSoldProperties:soldProperties,
-        totalFeaturedProperties:featuredProperties
+        totalRentedProperties: rentedProperties,
+        totalSoldProperties: soldProperties,
+        totalFeaturedProperties: featuredProperties,
+        banners: banners,
       },
     });
   } catch (error) {
@@ -76,8 +83,22 @@ export const deleteUser = async (req, res) => {
 //delete company
 export const deleteCompany = async (req, res) => {
   try {
+    const agentCompany = await AgentCompany.findById(req.params.id);
+    if (!agentCompany)
+      return res.status(404).json({ message: "agent company not found" });
+    await Agent.deleteMany({ companyId: agentCompany.companyId });
+    await SoldProperties.deleteMany({ companyId: agentCompany.companyId });
+    await Property.deleteMany({ companyId: agentCompany.companyId });
+    await PropertyAd.deleteMany({ companyId: agentCompany.companyId });
     await AgentCompany.findByIdAndDelete(req.params.id);
-    res.status(200).json("Agent Company deleted successfully");
+    await User.findOneAndUpdate(
+      { companyId: agentCompany.companyId },
+      { companyId: null, hasCompany: false, status: null },
+      {new:true}
+    );
+    res
+      .status(200)
+      .json("Agent Company and all related things are deleted successfully");
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -100,7 +121,9 @@ export const getAllProperties = async (req, res) => {
 
 export const getCompanyDetail = async (req, res) => {
   try {
-    const agentCompany = await AgentCompany.findOne({companyId: req.params.id})
+    const agentCompany = await AgentCompany.findOne({
+      companyId: req.params.id,
+    });
     const rentalProperties = await Property.find({
       companyId: req.params.id,
       type: "rent",
@@ -110,23 +133,37 @@ export const getCompanyDetail = async (req, res) => {
       type: "sell",
     }).count();
     const agents = await Agent.find({ companyId: req.params.id }).count();
-    const rentedProperty = await Property.find({companyId: req.params.id,isRented:true}).count();
-    const soldProperty = await Property.find({companyId: req.params.id,isSoldOut:true}).count();
-    const adProperty = await Property.find({companyId: req.params.id,isFeatured:true}).count();
-    const allProperties = await Property.find({companyId: req.params.id,}).sort({createdAt:-1})
-    const allAgents = await Agent.find({companyId: req.params.id,}).sort({createdAt:-1})
+    const rentedProperty = await Property.find({
+      companyId: req.params.id,
+      isRented: true,
+    }).count();
+    const soldProperty = await Property.find({
+      companyId: req.params.id,
+      isSoldOut: true,
+    }).count();
+    const adProperty = await Property.find({
+      companyId: req.params.id,
+      isFeatured: true,
+    }).count();
+    const allProperties = await Property.find({
+      companyId: req.params.id,
+    }).sort({ createdAt: -1 });
+    const allAgents = await Agent.find({ companyId: req.params.id }).sort({
+      createdAt: -1,
+    });
     res.status(200).json({
       success: true,
-      data: { 
-        owner:agentCompany,
-        rentProperties: rentalProperties, 
+      data: {
+        owner: agentCompany,
+        rentProperties: rentalProperties,
         sellProperties: sellProperties,
-         agents: agents,
-         properties:allProperties,
-         allAgents:allAgents,
-         soldProperty:soldProperty,
-         rentedProperty:rentedProperty,
-         adProperty:adProperty },
+        agents: agents,
+        properties: allProperties,
+        allAgents: allAgents,
+        soldProperty: soldProperty,
+        rentedProperty: rentedProperty,
+        adProperty: adProperty,
+      },
     });
   } catch (error) {
     res.status(500).json({ message: error.message });
